drm/msm: add timestamp param
We need this for GL_TIMESTAMP queries.

Note: currently only supported on a4xx.. a3xx doesn't have this
always-on counter.  I think we could emulate it with the one CP
counter that is available, but for now it is of limited usefulness
on a3xx (since we can't seem to do time-elapsed queries in any sane
way with the existing firmware on a3xx, and if you are trying to do
profiling on a tiler you want time-elapsed).  We can add that later
if it becomes useful.

Signed-off-by: Rob Clark <robdclark@gmail.com>

Conflicts:
	drivers/gpu/drm/msm/adreno/adreno_gpu.c
	include/uapi/drm/msm_drm.h

(cherry picked from commit 0b02ae1)
Signed-off-by: hexc0de <tsehcode@gmail.com>

Change-Id: I2c6d13e11a5d9c0883be1e70d95732000b161ab4