drm/msm: fix ->last_fence() after recover
It is no longer true that we discard all in-flight submits on recover
(these days we only discard the first one that hung).  After the first
re-submitted batch completes it would overwrite the fence with a correct
value, but there would be a window of time which showed all re-submitted
batches as already complete.

Signed-off-by: Rob Clark <robdclark@gmail.com>

(cherry picked from commit 54f1c8e)
Signed-off-by: hexc0de <tsehcode@gmail.com>

Change-Id: I66d4fb4c4ea41d82e2b46dce64f9ffd45fd648cc